requirements.txt
Django==5.0.2
django-allauth
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dj-rest-auth==5.0.2
django-cors-headers==4.3.1
celery==5.3.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
channels==4.0.0
daphne==4.1.0
redis==5.0.1
requests
pillow
django-filter
